+++ src/resources/ERScrollBox.cpp
@@ -3,27 +3,35 @@
 ERScrollBox::ERScrollBox(QWidget * parent)
 	: HorizontalScrollBox(parent)
 {
-	
 	// initialize menus
-	m_slide_menu = new QMenu(tr("ER context menu"), this);
-	m_action_new = new QAction("New Embedded Resource", m_slide_menu);
-	m_action_delete = new QAction("Delete Embedded Resource", m_slide_menu);
-	m_action_edit = new QAction("Edit Embedded Resource", m_slide_menu);
-	m_action_open = new QAction("Open Embedded Resource", m_slide_menu);
+	QMenu *m_menu = new QMenu(tr("ER menu"), this);
+	QMenu *m_item_menu = new QMenu(tr("ER item menu"), this);
 
-	m_slide_menu->addAction(m_action_new);
-	m_slide_menu->addAction(m_action_delete);
-	m_slide_menu->addAction(m_action_edit);
-	m_slide_menu->addAction(m_action_open);
+	QAction *m_action_new = new QAction("New Embedded Resource", this);
+	QAction *m_action_delete = new QAction("Delete Embedded Resource", this);
+	QAction *m_action_edit = new QAction("Edit Embedded Resource", this);
+	QAction *m_action_open = new QAction("Open Embedded Resource", this);
+	QAction *m_action_set = new QAction("Set Position", this);
+	QAction *m_action_goto = new QAction("Go to Position", this);
+
+	m_menu->addAction(m_action_new);
+	m_item_menu->addAction(m_action_new);
+	m_item_menu->addAction(m_action_delete);
+	m_item_menu->addAction(m_action_edit);
+	m_item_menu->addAction(m_action_open);
+	m_item_menu->addAction(m_action_set);
+	m_item_menu->addAction(m_action_goto);
 
 	// forward signals
 	connect(m_action_new, &QAction::triggered, this, &ERScrollBox::sNew);
 	connect(m_action_delete, &QAction::triggered, this, &ERScrollBox::sDelete);
 	connect(m_action_edit, &QAction::triggered, this, &ERScrollBox::sEdit);
 	connect(m_action_open, &QAction::triggered, this, &ERScrollBox::sOpen);
+	connect(m_action_set, &QAction::triggered, this, &ERScrollBox::sSetPosition);
+	connect(m_action_goto, &QAction::triggered, this, &ERScrollBox::sGotoPosition);
 
-	setMenu(m_slide_menu);
-	setItemMenu(m_slide_menu);
+	setMenu(m_menu);
+	setItemMenu(m_item_menu);
 }
 
 ERScrollBox::~ERScrollBox() {