+++ src/resources/ERScrollBox.h
@@ -19,13 +19,8 @@ public:
 signals:
 	void sEdit(); // only support single edit
 	void sOpen();
-
-private:
-	QMenu* m_slide_menu; // context menu
-	QAction* m_action_new;
-	QAction* m_action_delete;
-	QAction* m_action_edit;
-	QAction* m_action_open;
+	void sSetPosition();
+	void sGotoPosition();
 
 	//int filter; // 0 global, 1 local, 2 
 };